Consolidate caching tests
